+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/sun/FlatCorrection.java
@@ -17,29 +17,46 @@ package me.champeau.a4j.jsolex.processing.sun;
 
 import me.champeau.a4j.jsolex.processing.util.Histogram;
 import me.champeau.a4j.jsolex.processing.util.ImageWrapper32;
+import me.champeau.a4j.math.image.ImageMath;
+import me.champeau.a4j.math.image.Kernel33;
 import me.champeau.a4j.math.regression.Ellipse;
+import org.apache.commons.math3.analysis.polynomials.PolynomialFunction;
+import org.apache.commons.math3.fitting.PolynomialCurveFitter;
+import org.apache.commons.math3.fitting.WeightedObservedPoints;
 
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.stream.IntStream;
 
 public class FlatCorrection {
-    private static final int MARGIN = 5;
-    private static final int BAND_SIZE = 1;
+    public static final double DEFAULT_LO_PERCENTILE = 0.1;
+    public static final double DEFAULT_HI_PERCENTILE = 0.95;
+    public static final int DEFAULT_ORDER = 2;
 
-    private FlatCorrection() {
+    private final double loPercentile;
+    private final double hiPercentile;
+    private final int order;
 
+    public FlatCorrection() {
+        this(DEFAULT_LO_PERCENTILE, DEFAULT_HI_PERCENTILE, DEFAULT_ORDER);
     }
 
-    public static double[] computeCorrectionFactors(ImageWrapper32 source, Ellipse ellipse) {
-        int width = source.width();
-        int height = source.height();
-        var data = source.data();
+    public FlatCorrection(double loPercentile, double hiPercentile, int order) {
+        this.loPercentile = loPercentile;
+        this.hiPercentile = hiPercentile;
+        this.order = order;
+    }
+
+    public double[] computeCorrectionFactors(ImageWrapper32 source, Ellipse ellipse) {
+        var blurred = ImageMath.newInstance().convolve(source.asImage(), Kernel33.GAUSSIAN_BLUR);
+        int width = blurred.width();
+        int height = blurred.height();
+        var data = blurred.data();
         var minX = width;
         var maxX = 0;
         var minY = height;
         var maxY = 0;
-        var maxValue = 0d;
+        var maxValue = -Double.MAX_VALUE;
         var minValue = Double.MAX_VALUE;
         var builder = Histogram.builder(65536);
         for (int y = 0; y < height; y++) {
@@ -57,50 +74,61 @@ public class FlatCorrection {
             }
         }
         var diskHistogram = builder.build().cumulative();
-        var percentile = diskHistogram.percentile(0.95);
-        var lo = 0.05 * percentile;
-        var hi = diskHistogram.percentile(0.99);
-        var medianValues = new double[height];
+        var lo = diskHistogram.percentile(loPercentile);
+        var hi = diskHistogram.percentile(hiPercentile);
+        var avgValues = new double[height];
         IntStream.range(0, height).parallel().forEach(y -> {
-            var builder2 = Histogram.builder(65536);
+            double total = 0;
+            double count = 0;
             for (int x = 0; x < width; x++) {
-                for (int y1 = Math.max(0, y - BAND_SIZE); y1 < Math.min(height, y + BAND_SIZE); y1++) {
-                    if (ellipse.isWithin(x, y1)) {
-                        var v = data[x + y1 * width];
-                        if (v > lo && v < hi) {
-                            builder2.record(v);
-                        }
+                if (ellipse.isWithin(x, y)) {
+                    var v = data[x + y * width];
+                    if (v > lo && v < hi) {
+                        total += v;
+                        count++;
                     }
                 }
             }
-            var histogram = builder2.build().cumulative();
-            var median = Histogram.medianOf(histogram);
-            medianValues[y] = median;
+            avgValues[y] = count == 0 ? 0 : total / count;
         });
-        var total = 0d;
-        for (int k = 0; k < MARGIN; k++) {
-            total += medianValues[minY + k];
-        }
-        for (int y = 0; y < minY + MARGIN; y++) {
-            medianValues[y] = total / MARGIN;
-        }
-        total = 0d;
-        for (int k = 0; k < MARGIN; k++) {
-            total += medianValues[maxY - k];
-        }
-        for (int y = maxY - MARGIN + 1; y < height; y++) {
-            medianValues[y] = total / MARGIN;
-        }
-        var avg = Arrays.stream(medianValues).average().orElse(1e-7);
+        var weights = new double[height];
+        // iterate over median values. If a value is 0, then replace it with the closest non-zero value
         for (int y = 0; y < height; y++) {
-            medianValues[y] = Math.max(avg, medianValues[y]);
+            if (avgValues[y] < lo) {
+                weights[y] = 1;
+                int left = y;
+                while (left > 0 && avgValues[left] < lo) {
+                    left--;
+                }
+                int right = y;
+                while (right < height - 1 && avgValues[right] < lo) {
+                    right++;
+                }
+                if (avgValues[left] > lo && avgValues[right] > lo) {
+                    avgValues[y] = (avgValues[left] + avgValues[right]) / 2;
+                } else if (avgValues[left] > lo) {
+                    avgValues[y] = avgValues[left];
+                } else if (avgValues[right] > lo) {
+                    avgValues[y] = avgValues[right];
+                }
+            } else {
+                weights[y] = 4;
+            }
         }
-        var smoothed = performSmoothing(medianValues);
-        var max = Arrays.stream(smoothed).max().orElse(1e-7);
-        return Arrays.stream(smoothed).map(v -> v / max).toArray();
+        // perform polynomial regression on the median values
+        var regression = PolynomialCurveFitter.create(order);
+        var series = new WeightedObservedPoints();
+        for (int y = 0; y < height; y++) {
+            series.add(weights[y], y, avgValues[y]);
+        }
+        var fit = regression.fit(series.toList());
+        var fun = new PolynomialFunction(fit);
+        var fitted = IntStream.range(0, height).mapToDouble(fun::value).toArray();
+        var max = Arrays.stream(fitted).max().orElse(1e-7);
+        return Arrays.stream(fitted).map(v -> v / max).toArray();
     }
 
-    public static ImageWrapper32 correctImage(ImageWrapper32 source, double[] normalized) {
+    public ImageWrapper32 correctImage(ImageWrapper32 source, double[] normalized) {
         int width = source.width();
         int height = source.height();
         var data = source.data();
@@ -115,15 +143,10 @@ public class FlatCorrection {
         return new ImageWrapper32(width, height, outData, new HashMap<>(source.metadata()));
     }
 
-    private static double[] performSmoothing(double[] values) {
-        double[] result = new double[values.length];
-        double max = 0;
-        result[0] = (2 * values[0] + values[1]) / 3d;
-        for (int i = 1; i < values.length - 1; i++) {
-            result[i] = (values[i - 1] + 2 * values[i] + values[i + 1]) / 4d;
-            max = Math.max(max, result[i]);
-        }
-        result[values.length - 1] = (2 * values[values.length - 1] + values[values.length - 2]) / 3d;
-        return result;
+    public ImageWrapper32 correctImage(ImageWrapper32 source) {
+        return source.findMetadata(Ellipse.class)
+            .map(ellipse -> correctImage(source, computeCorrectionFactors(source, ellipse)))
+            .orElse(source);
     }
+
 }